Initial commit of new trial application
This commit will demonstrate how to easily pull Spring Boot 2.7
and related Spring artifacts from the HeroDevs NES registry.
The sample applciation is replicated two times, one for Maven
and one for Gradle. This is to prioritize user experience
above code duplication. The sample application in both folders
is identical.